Fix the landscape mode appearance of the recorder screen
