+++ app/build.gradle
@@ -65,6 +65,9 @@ dependencies {
     implementation 'androidx.compose.material3:material3:1.1.1'
     implementation 'androidx.compose.material:material-icons-extended:1.4.3'
 
+    //Constraint Layout
+    implementation "androidx.constraintlayout:constraintlayout-compose:1.0.1"
+
     //Navigation
     implementation 'androidx.navigation:navigation-compose:2.6.0'
